study-resource-sharing-platform/src/test/java/com/studyshare/platform/StudyResourceSharingPlatformApplicationTests.java
package com.studyshare.platform;

import org.junit.jupiter.api.Test;
import org.springframework.boot.test.context.SpringBootTest;
import static org.junit.jupiter.api.Assertions.assertTrue;

/**
 * Unit test class for {@link StudyResourceSharingPlatformApplication}.
 * <p>
 * This class contains basic tests to verify that the application context loads correctly.
 * </p>
 *
 * @see StudyResourceSharingPlatformApplication
 * @since 1.0
 */
@SpringBootTest
class StudyResourceSharingPlatformApplicationTests {

    /**
     * Verifies that the Spring application context loads successfully.
     * <p>
     * This test is a basic "smoke test" to ensure that the Spring Boot application
     * configuration is valid and that the application can start without issues.
     * </p>
     */
    @Test
    void contextLoads() {
        // Basic context load test
        assertTrue(true);
    }
}
